pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pdb_align"
version = "0.1.0"
description = "High-performance protein structure alignment package."
dependencies = [
    "numpy<2.0", "pandas", "scipy", "scikit-learn", "biopython", "numba", "requests"
]

[project.optional-dependencies]
app = [
    "streamlit", "py3Dmol", "stmol", "plotly", "matplotlib", "seaborn", "streamlit-molstar"
]
